fix(ui): address PR review — URL API, deps, codecov, test layout
Use URL() for request resolution, simplify readErrorBody, bump
react-router-dom to 7.9.0, consolidate tests/fixtures and tests/utils,
upload ui lcov to Codecov in CI, and document apiFetch/AuthGuard patterns.
