+++ ui/package.json
@@ -19,7 +19,7 @@
   "dependencies": {
     "react": "19.2.6",
     "react-dom": "19.2.6",
-    "react-router-dom": "7.8.0"
+    "react-router-dom": "7.9.0"
   },
   "devDependencies": {
     "@eslint/js": "9.39.4",
